pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "openai>=2.8.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
